requirements.txt
# Core Streamlit
streamlit>=1.30.0

# LangChain Core
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-google-genai>=2.0.0

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# ML & Data Processing
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Environment & Config
python-dotenv>=1.0.0

# Google AI
google-generativeai>=0.8.0

# Evaluation
scikit-learn>=1.3.0
